Добавил простую валидацию на стороне сервера
В дальнейшем заменю валидацию на более сложную, пока в этом нет необходимости.
Исправил баг который возникал после возникновения ошибки о неправильном вводе данных.

+++ src/main/java/com/webforms/controllers/ApiController.java
@@ -15,18 +15,27 @@ public class ApiController {
     @PostMapping(path = "/add", consumes = "application/json")
     public @ResponseBody String add(@RequestBody Item item){
         //System.out.println(item);
+        if(!item.isValid()){
+            return "Not Saved";
+        }
         userRepository.save(item);
         return "Saved";
     }
 
     @PostMapping(path = "/delete")
     public @ResponseBody String delete(@RequestBody Item item){
+        if(!item.isValid()){
+            return "Not Deleted";
+        }
         userRepository.delete(item);
         return "Deleted";
     }
 
     @PostMapping(path = "/update")
     public @ResponseBody String update(@RequestBody Item item){
+        if(!item.isValid()){
+            return "Not Updated";
+        }
         userRepository.deleteById(item.getId());
         userRepository.save(item);
         return "Updated";
